requirements.txt
PyMuPDF
python-pptx
flask
huggingface_hub